app/requirements.txt
flask
simplejson
image
python-dotenv